requirements.txt
djangorestframework==3.16.0
psycopg2-binary==2.9.10
pytz==2025.2
